[Feat] add query

+++ backend/src/main/java/com/hanati/team1/src/players/models/GetProspectRes.java
@@ -3,7 +3,7 @@ package com.hanati.team1.src.players.models;
 public record GetProspectRes(
 	long id,
 	String name,
-	int cntSub,
+	long cntSub,
 	String profile,
 	String position
 ) {
